Internal Interface

The internal interface for UMAP.jl. Some of these may be moved to the public interface as they stabilize.

Contents

Internal Interface

Contents

Index

Interface

Index

UMAP.DescentNeighbors

UMAP.NeighborParams

UMAP.PrecomputedNeighbors

UMAP.SourceGlobalParams

UMAP.SourceViewParams

UMAP._knn_from_dists

UMAP.coalesce_views

UMAP.fuzzy_simplicial_set

UMAP.fuzzy_simplicial_set

UMAP.fuzzy_simplicial_set

UMAP.fuzzy_simplicial_set

UMAP.knn_search

UMAP.knn_search

UMAP.knn_search

UMAP.knn_search

UMAP.knn_search

UMAP.knn_search

UMAP.knn_search

UMAP.knn_search

Interface

Nearest Neighbors Interface

UMAP.jl allows for an extensible way to find the nearest neighbors of your data. This is controlled by the subtypes of NeighborParams.

UMAP.NeighborParams — Type

Structs for parameterizing the knn search step of UMAP.

Subtyping NeighborParams allows for different methods of finding nearest neighbors. The knn_search function will dispatch on the type of knn_params to find the nearest neighbors in the data.

UMAP.DescentNeighbors — Type

DescentNeighbors(n_neighbors, metric, kwargs)

Parameters for finding approximate nearest neighbors using NearestNeighborDescent.

UMAP.PrecomputedNeighbors — Type

PrecomputedNeighbors(n_neighbors, dists_or_graphs)

Parameters for finding nearest neighbors from precomputed distances. dists_or_graphs can either be a matrix (pairwise distances) or a NearestNeighborDescent.ApproximateKNNGraph.

The neighbor parameters control dispatch for KNN search functionality, both when fitting and when transforming new data.

UMAP.knn_search — Function

knn_search(data, knn_params)
knn_search(data, queries, knn_params, results)

Find the (potentially approximate) k-nearest neighbors for each point in the dataset, according to knnparams. The data may consist of one or more 'views', passed in either directly (for a single view only) or as a NamedTuple, with corresponding knnparams (either an instance of <: NeighborParams or a NamedTuple of such parameters).

For each (dataview, knnparams) pair, two KxN matrices are returned (knns, dists), holding the indices of the k-nearest neighbors and distances to those neighbors for each point in the dataset. When knnparams isa DescentNeighbors, these matrices are computed using NearestNeighborDescent. When knnparams isa PrecomputedNeighbors, the knn matrices are created from user-provided distances.

If data and knnparams are NamedTuples, the returned knnmatrices will also be in a NamedTuple with the same keys as the inputs.

UMAP.knn_search — Method

knn_search(data::NamedTuple{T}, knn_params::NamedTuple{T}) -> NamedTuple{T}

(FIT) Map knn_search over each view of the data and corresponding knn_params.

UMAP.knn_search — Method

knn_search(data::NamedTuple{T}, queries::NamedTuple{T}, knn_params::NamedTuple{T}, result_knns_dists::NamedTuple{T}) -> NamedTuple{T}

(TRANSFORM) Map knn_search over each view of the data, queries, knnparams, and resultknns_dists.

UMAP.knn_search — Method

knn_search(data, knn_params::DescentNeighbors) -> (knns, dists)

Find approximate nearest neighbors using nndescent.

UMAP.knn_search — Method

knn_search(data, queries, knn_params::DescentNeighbors, result_knns_dists) -> (knns, dists)

Search for approximate nearest neighbors of queries in data using nndescent. The (knns, dists) are used to reconstruct the original KNN graph.

UMAP.knn_search — Method

knn_search(_, knn_params::PrecomputedNeighbors{M}) where {M <: ApproximateKNNGraph}

Get neighbors from a precomputed KNNGraph. This method is used when the KNN graph has been precomputed using NearestNeighborDescent

UMAP.knn_search — Method

knn_search(_, knn_params::PrecomputedNeighbors)

(FIT) Get neighbors from a precomputed distance matrix.

UMAP.knn_search — Method

knn_search(_, __, knn_params::PrecomputedNeighbors, ___)

(TRANSFORM) Get neighbors from a precomputed distance matrix.

UMAP._knn_from_dists — Function

_knn_from_dists(dist_mat::AbstractMatrix{S}, k::Integer; ignore_diagonal=true) where {S <: Real}

Construct k-nearest neighbors and distances from a distance matrix. If ignore_diagonal is true, the diagonal elements (which are 0) will be ignored when determining the k-nearest neighbors. This is useful when the distance matrix represents pairwise distances of the same set, where the diagonal elements are the distances to themselves and will always be 0.

Returns a tuple of two matrices: the first contains the indices of the k-nearest neighbors, and the second contains the corresponding distances to those neighbors.

Fuzzy Simplicial Sets Interface

The second step of the UMAP algorithm involves calculating a topological representation of the data (potentially from multiple "views"). This behavior is controlled by the following types and functions.

UMAP.SourceViewParams — Type

SourceViewParams(set_operation_ratio, local_connectivity, bandwidth)

Struct for parameterizing the representation of the data in the source (original) manifold; i.e. constructing fuzzy simplicial sets of each view of the dataset.

UMAP.SourceGlobalParams — Type

SourceGlobalParams{T}(mix_ratio)

Parameters for merging the fuzzy simplicial sets for each dataset view into one fuzzy simplicial set, otherwise known as the UMAP graph.

These config structs parameterize the fuzzy simplicial sets created for each view, and the global fuzzy simplicial set after combining all views (sometimes called the UMAP graph).

UMAP.fuzzy_simplicial_set — Function

Construct the UMAP graph, i.e. the global fuzzy simplicial set. This is represented as a symmetric, sparse matrix where each value is the probability that an edge exists between points (row, col).

For multiple views, there will be one UMAP graph per view - later combined via coalesce_views.

For transforming new data, this graph is notably not symmetric.

UMAP.fuzzy_simplicial_set — Method

fuzzy_simplicial_set(knns_dists::NamedTuple{T}, knn_params::NamedTuple{T}, src_params::NamedTuple{T}) -> NamedTuple{T}

(Fit) Construct a fuzzy simplicial set for each view of the data, returning a named tuple of fuzzy simplicial sets, one for each view. The keys of the named tuple are the keys of the input NamedTuples.

UMAP.fuzzy_simplicial_set — Method

fuzzy_simplicial_set(data::NamedTuple{T}, knns_dists::NamedTuple{T}, knn_params::NamedTuple{T}, src_params::NamedTuple{T}) -> NamedTuple{T}

(Transform) Construct a fuzzy simplicial set for each view of the data, returning a named tuple of fuzzy simplicial sets, one for each view.

UMAP.fuzzy_simplicial_set — Method

fuzzy_simplicial_set(knns_dists, n_points, knn_params, src_params, combine=true) -> membership_graph::SparseMatrixCSC

Construct the local fuzzy simplicial set of each point represented by its distances to its nearest neighbors, stored in knns and dists, normalizing the distances, and converting the metric space to a simplicial set (a weighted graph).

n_points indicates the total number of points of the original data, while knns contains indices of some subset of those points (ie some subset of 1:n_points). If knns represents neighbors of the elements of some set with itself, then knns should have n_points number of columns. Otherwise, these two values may be different.

If combine is true, use intersections and unions to combine local fuzzy sets of neighbors. The returned graph has size (n_points, size(knns, 2)).

UMAP optimizes the difference between the source fuzzy set representation and a target representation. Therefore, if there are multiple source views, they need to be combined. This is done by a function coalesce_views that combines the view simplicial sets via set union and intersections, controlled by the SourceGlobalParams.

UMAP.coalesce_views — Function

coalesce_views(view_fuzzy_sets, params)

Merge the fuzzy simplicial sets for each view of the data. This returns a single fuzzy simplicial set - the weighted, undirected UMAP graph - that captures the spatial relationships of the data points approximated by the manifold.